Share service client with dev web bridge
